backend/requirements.txt
fastapi
uvicorn
torch
transformers
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
mlflow